Modificados resources

+++ src/main/java/es/smartweekend/web/backend/jersey/resources/NewsResource.java
@@ -1,5 +1,6 @@
 package es.smartweekend.web.backend.jersey.resources;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.ws.rs.Consumes;
@@ -17,17 +18,30 @@ import javax.ws.rs.core.Response;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import es.smartweekend.web.backend.jersey.util.ApplicationContextProvider;
 import es.smartweekend.web.backend.model.newsItem.NewsItem;
 import es.smartweekend.web.backend.model.newsService.NewsService;
 import es.smartweekend.web.backend.model.util.exceptions.ServiceException;
+import es.smartweekend.web.backend.model.util.session.SessionManager;
 
 /**
  * @author Miguel Ángel Castillo Bellagona
  */
+@Path("news")
 public class NewsResource {
 
+	private String[] s = {"newsItemId","title","creationDate","publishDate","content","publisher","publisher.login","event"};
+	private ArrayList<String> l;
+	
 	@Autowired
 	private NewsService newsService;
+	
+	public NewsResource() {
+		this.newsService  = ApplicationContextProvider.getApplicationContext().getBean(NewsService.class);
+		l = new ArrayList<String>();
+		l.add(s[0]);l.add(s[1]);l.add(s[2]);l.add(s[3]);l.add(s[4]);l.add(s[5]);l.add(s[6]);l.add(s[7]);
+	}
+	
 	
 	//ANONYMOUS
 	
@@ -45,6 +59,7 @@ public class NewsResource {
 		boolean b = true;
 		if(desc==0) b = false;
 		try {
+			if(l.indexOf(orderBy)<0) throw new ServiceException(ServiceException.INCORRECT_FIELD,"orderBy");
 			List<NewsItem> l =  newsService.getPublishedNewsItemFromEvent(eventId,startIndex,cont,orderBy,b);
 			return Response.status(200).entity(l).build();
 		} catch (ServiceException e) {
@@ -75,6 +90,7 @@ public class NewsResource {
 	@Produces({MediaType.APPLICATION_JSON})
 	public Response NewsADMIN(@HeaderParam("sessionId") String sessionId, NewsItem newsItem) {
 		try {
+			if(!SessionManager.exists(sessionId)) throw new ServiceException(ServiceException.INVALID_SESSION);
 			NewsItem n = newsService.addNewsADMIN(sessionId, newsItem);
 			return Response.status(200).entity(n).build();
 		} catch (ServiceException e) {
@@ -88,6 +104,7 @@ public class NewsResource {
 	@Produces({MediaType.APPLICATION_JSON})
 	public Response changeNewsDataADMIN(@HeaderParam("sessionId") String sessionId, @PathParam("newsItemId") int newsItemId, NewsItem newsData) {
 		try {
+			if(!SessionManager.exists(sessionId)) throw new ServiceException(ServiceException.INVALID_SESSION);
 			NewsItem n = newsService.changeNewsDataADMIN(sessionId, newsItemId, newsData);
 			return Response.status(200).entity(n).build();
 		} catch (ServiceException e) {
@@ -100,6 +117,7 @@ public class NewsResource {
 	@Produces({MediaType.APPLICATION_JSON})
 	public Response getNewsItemADMIN(@HeaderParam("sessionId") String sessionId, @PathParam("newsItemId") int newsItemId) {
 		try {
+			if(!SessionManager.exists(sessionId)) throw new ServiceException(ServiceException.INVALID_SESSION);
 			NewsItem n =  newsService.getNewsItemADMIN(sessionId, newsItemId);
 			return Response.status(200).entity(n).build();
 		} catch (ServiceException e) {
@@ -122,6 +140,8 @@ public class NewsResource {
 		boolean b = true;
 		if(desc==0) b = false;
 		try {
+			if(!SessionManager.exists(sessionId)) throw new ServiceException(ServiceException.INVALID_SESSION);
+			if(l.indexOf(orderBy)<0) throw new ServiceException(ServiceException.INCORRECT_FIELD,"orderBy");
 			List<NewsItem> l = newsService.getAllNewsItemADMINFromEvent(sessionId,eventId,startIndex,cont,orderBy,b);
 			return Response.status(200).entity(l).build();
 		} catch (ServiceException e) {
@@ -134,6 +154,7 @@ public class NewsResource {
 	@Produces({MediaType.APPLICATION_JSON})
 	public Response getAllNewsItemTamADMIN(@HeaderParam("sessionId") String sessionId, @PathParam("eventId") int eventId) {
 		try {
+			if(!SessionManager.exists(sessionId)) throw new ServiceException(ServiceException.INVALID_SESSION);
 			long l = newsService.getAllNewsItemTamADMINFromEvent(sessionId,eventId);
 			return Response.status(200).entity(l).build();
 		} catch (ServiceException e) {
@@ -145,6 +166,7 @@ public class NewsResource {
 	@DELETE
     public Response removeNewsADMIN(@HeaderParam("sessionId") String sessionId, @PathParam("newsItemId") int newsItemId) {
 		try {
+			if(!SessionManager.exists(sessionId)) throw new ServiceException(ServiceException.INVALID_SESSION);
 			newsService.removeNewsADMIN(sessionId, newsItemId);
 			return Response.status(203).build();
 		} catch (ServiceException e) {
